VxScan: export CDF CVRs by default
