+++ apps/scan/backend/jest.config.js
@@ -13,9 +13,9 @@ module.exports = {
   collectCoverageFrom: ['**/*.{ts,tsx}', '!**/node_modules/**', '!test/**/*'],
   coverageThreshold: {
     global: {
-      statements: 89,
+      statements: 88,
       branches: 75,
-      functions: 90,
+      functions: 88,
       lines: 89,
     },
   },
