+++ apps/scan/backend/src/env.d.ts
@@ -8,7 +8,6 @@ declare namespace NodeJS {
     readonly SCAN_WORKSPACE?: string;
     readonly VX_MACHINE_ID?: string;
     readonly VX_CODE_VERSION?: string;
-    readonly CVR_EXPORT_FORMAT?: 'cdf' | 'vxf';
     readonly SCANNER_MODEL?: 'custom' | 'plustek';
   }
 }
