+++ apps/scan/backend/src/tally-cvrs/validation.test.ts
