+++ apps/scan/backend/src/tally-cvrs/validation.ts
